+++ lib/http_server_fixture/data/test-data.json
@@ -1,1 +1,1 @@
-{"1":{"task":"a new task"},"2":{"task":"a new task"}}
+{"1":{"task":"a new task"},"2":{"task":"a new task"},"3":{"task":"a new task"}}
